Chromium: 6e3a05d

+++ src/base/metrics/sample_map.cc
@@ -19,7 +19,7 @@ SampleMap::~SampleMap() {}
 
 void SampleMap::Accumulate(Sample value, Count count) {
   sample_counts_[value] += count;
-  IncreaseSum(count * value);
+  IncreaseSum(static_cast<int64_t>(count) * value);
   IncreaseRedundantCount(count);
 }
